+++ network_protocol/src/model/header.rs
@@ -1,0 +1,56 @@
+use crate::{MAX_ID_LEN, MAX_ID_MESS_LEN, MAX_SEQ_NUMBER_LEN, MessageCreationError};
+
+/// This is only for definition and fields should never be accessed nor modified directly
+/// use the methods provided with the struct instead
+pub struct Header {
+    id_dest: u8, // really a u4
+    id_src: u8, // really a u4
+    is_ack: bool, // really a u1
+    id_message: u8, // really a u3
+    seq_number: u8, // really a u4
+}
+
+impl Header {
+    /// Creates a new header with the given parameters
+    /// # Warnining
+    /// take care of the size of the parameters, the function will return an error if the size is not correct
+    /// # Arguments
+    /// id_dest: u4,
+    /// id_src: u4,
+    /// is_ack: u1,
+    /// id_message: u3,
+    /// seq_number: u4
+    pub fn new(id_dest: u8, id_src: u8, is_ack: bool, id_message: u8, seq_number: u8) -> Result<Header,MessageCreationError> {
+
+        if id_dest > MAX_ID_LEN || id_src > MAX_ID_LEN || id_message > MAX_ID_MESS_LEN || seq_number > MAX_SEQ_NUMBER_LEN {
+            return Err(MessageCreationError::ParametersTooLong);
+        }
+        if id_dest == id_src {
+            return Err(MessageCreationError::SrcAndDestCanNotBeEqual);
+        }
+
+        Ok(Header {
+            id_dest,
+            id_src,
+            is_ack,
+            id_message,
+            seq_number,
+        })
+    }
+
+    pub fn get_id_dest(&self) -> u8 {
+        self.id_dest
+    }
+    pub fn get_id_src(&self) -> u8 {
+        self.id_src
+    }
+    pub fn get_is_ack(&self) -> bool {
+        self.is_ack
+    }
+    pub fn get_id_message(&self) -> u8 {
+        self.id_message
+    }
+    pub fn get_seq_number(&self) -> u8 {
+        self.seq_number
+    }
+}
