+++ network_protocol/src/model/protocol_constants.rs
@@ -1,0 +1,5 @@
+
+pub const MAX_ID_LEN: u8 = 16 -1;
+pub const MAX_SEQ_NUMBER_LEN: u8 = 16 -1;
+pub const MAX_ID_MESS_LEN: u8 = 8 -1;
+pub const BYTES_LEFT_IN_PACKAGE: u8 = 6;
